requirements.txt
# Web Framework & Server
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20

# Database & ORM
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.11.9
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.1.1
PyYAML==6.0.3

# HTTP & API Communication
requests==2.32.5
httpx==0.27.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.1

# Async & Concurrency
anyio==4.10.0
sniffio==1.3.1

# Development & Testing
pytest==8.4.2
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
typing-extensions==4.15.0
